auth_jwt

+++ src/main/java/com/springboot/project/airBnbApp/service/impl/HotelServiceImpl.java
@@ -5,7 +5,9 @@ import com.springboot.project.airBnbApp.dto.HotelInfoDto;
 import com.springboot.project.airBnbApp.dto.RoomDto;
 import com.springboot.project.airBnbApp.entity.Hotel;
 import com.springboot.project.airBnbApp.entity.Room;
+import com.springboot.project.airBnbApp.entity.User;
 import com.springboot.project.airBnbApp.exception.ResourceNotFoundException;
+import com.springboot.project.airBnbApp.exception.UnAuthorisedException;
 import com.springboot.project.airBnbApp.repository.HotelRepository;
 import com.springboot.project.airBnbApp.repository.RoomRepository;
 import com.springboot.project.airBnbApp.service.HotelService;
@@ -13,6 +15,7 @@ import com.springboot.project.airBnbApp.service.InventoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,6 +37,10 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating a new hotel with name : {}",hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with id : {}",hotelDto.getId());
         return modelMapper.map(hotel,HotelDto.class);
@@ -43,6 +50,12 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto getHotelById(Long id) {
         log.info("Getting hotel with id : {}",id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with id: " +id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id : "+id);
+        }
+
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -50,6 +63,12 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto updateHotelById(Long id, HotelDto hotelDto) {
         log.info("Updating hotel with id : {}",id);
         Hotel hotelToUpdate = hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with id :"+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotelToUpdate.getOwner())){
+            throw new UnAuthorisedException("User does not own this hotel :"+id);
+        }
+
         modelMapper.map(hotelDto,hotelToUpdate);
         Hotel hotel = hotelRepository.save(hotelToUpdate);
         return modelMapper.map(hotel,HotelDto.class);
@@ -61,6 +80,11 @@ public class HotelServiceImpl implements HotelService {
     public Boolean deleteHotelById(Long id) {
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with id: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("User does not own this hotel :"+id);
+        }
+
         for(Room room : hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -76,6 +100,12 @@ public class HotelServiceImpl implements HotelService {
         log.info("Activating the hotel : {}",hotelId);
         Hotel hotel = hotelRepository.findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with id: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("User does not own this hotel :"+id);
+        }
+
         if(hotel.getActive()==false) {
             hotel.setActive(true);
             for(Room room : hotel.getRooms()){
